# docs/admin/openapi.json
{
  "components": {
    "schemas": {
...
      "AnnotationOut": {
        "properties": {
          "ai_description": {
            "title": "Ai Description",
            "type": "string"
          },
          "ai_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Model"
          },
          "ai_prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Prompt Version"
          },
          "ai_status": {
            "enum": [
              "real",
              "fallback",
              "failed",
              "none"
            ],
            "title": "Ai Status",
            "type": "string"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "drop_clip": {
            "title": "Drop Clip",
            "type": "boolean"
          },
          "event_id": {
            "title": "Event Id",
            "type": "integer"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "frame_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Count"
          },
          "frame_size": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Size"
          },
          "needs_review": {
            "title": "Needs Review",
            "type": "boolean"
          },
          "preload_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preload Source"
          },
          "review_frame": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Frame"
          },
          "review_note": {
            "default": "",
            "title": "Review Note",
            "type": "string"
          },
          "status": {
            "enum": [
              "new",
              "edited",
              "submitted",
              "reviewed",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "suggestions_used": {
            "title": "Suggestions Used",
            "type": "boolean"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "title": "Tracks",
            "type": "array"
          },
          "updated_utc": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated Utc"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "event_id",
          "version",
          "status",
          "tracks",
          "description",
          "ai_description",
          "ai_status",
          "ai_model",
          "ai_prompt_version",
          "drop_clip",
          "needs_review",
          "author",
          "updated_utc",
          "fps",
          "frame_count",
          "frame_size",
          "suggestions_used"
        ],
        "title": "AnnotationOut",
        "type": "object"
      },
...
      "ClipAnnotationIn": {
        "properties": {
          "base_version": {
            "title": "Base Version",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "drop_clip": {
            "default": false,
            "title": "Drop Clip",
            "type": "boolean"
          },
          "key": {
            "maxLength": 512,
            "title": "Key",
            "type": "string"
          },
          "needs_review": {
            "default": false,
            "title": "Needs Review",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "edited",
              "submitted"
            ],
            "title": "Status",
            "type": "string"
          },
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/Track"
            },
            "title": "Tracks",
            "type": "array"
          }
        },
        "required": [
          "base_version",
          "tracks",
          "description",
          "status",
          "key"
        ],
        "title": "ClipAnnotationIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TagConversion": {
        "description": "\"In my words\" restructured into the clip's answer schema (tagstudio/convert.py): a suggestion for the form.",
        "properties": {
          "fields": {
            "additionalProperties": true,
            "title": "Fields",
            "type": "object"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "raw": {
            "title": "Raw",
            "type": "string"
          },
          "schema_name": {
            "title": "Schema Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "model",
          "language",
          "prompt_version",
          "schema_name",
          "fields",
          "raw"
        ],
        "title": "TagConversion",
        "type": "object"
      },
      "TagConvertRequest": {
        "properties": {
          "key": {
            "maxLength": 512,
            "title": "Key",
            "type": "string"
          },
          "words": {
            "maxLength": 2000,
            "title": "Words",
            "type": "string"
          }
        },
        "required": [
          "key",
          "words"
        ],
        "title": "TagConvertRequest",
        "type": "object"
      },
...
      "TaggingClip": {
        "properties": {
          "answer_schema": {
            "additionalProperties": true,
            "default": {},
            "title": "Answer Schema",
            "type": "object"
          },
          "assessment": {
            "additionalProperties": true,
            "title": "Assessment",
            "type": "object"
          },
          "form": {
            "additionalProperties": true,
            "title": "Form",
            "type": "object"
          },
          "fps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "History",
            "type": "array"
          },
          "item": {
            "additionalProperties": true,
            "title": "Item",
            "type": "object"
          },
          "media": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Media",
            "type": "object"
          },
          "media_reasons": {
            "additionalProperties": {
              "type": "string"
            },
            "default": {},
            "title": "Media Reasons",
            "type": "object"
          },
          "opinions": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Opinions",
            "type": "object"
          },
          "prefilled_from": {
            "title": "Prefilled From",
            "type": "string"
          },
          "prompt_version": {
            "default": "",
            "title": "Prompt Version",
            "type": "string"
          },
          "tag": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          }
        },
        "required": [
          "item",
          "media",
          "opinions",
          "form",
          "prefilled_from",
          "assessment",
          "history"
        ],
        "title": "TaggingClip",
        "type": "object"
      },
...
      "TaggingExportRequest": {
        "properties": {
          "include_needs_check": {
            "default": false,
            "title": "Include Needs Check",
            "type": "boolean"
          }
        },
        "title": "TaggingExportRequest",
        "type": "object"
      },
...
      "TaggingMediaRequest": {
        "properties": {
          "key": {
            "maxLength": 512,
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "clip",
              "crop"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "key",
          "kind"
        ],
        "title": "TaggingMediaRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Home Guard Admin Center",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tagging/boxes": {
      "get": {
        "operationId": "tagging_boxes_v1_tagging_boxes_get",
        "parameters": [
          {
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "maxLength": 512,
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Boxes",
        "tags": [
          "tagging"
        ]
      },
      "put": {
        "operationId": "tagging_save_boxes_v1_tagging_boxes_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClipAnnotationIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Save Boxes",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/clip": {
      "get": {
        "operationId": "tagging_clip_v1_tagging_clip_get",
        "parameters": [
          {
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "maxLength": 512,
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaggingClip"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Clip",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/convert": {
      "post": {
        "operationId": "tagging_convert_v1_tagging_convert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagConvertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagConversion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Convert",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/export": {
      "post": {
        "operationId": "tagging_export_v1_tagging_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaggingExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaggingExportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Export",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/file/{token}": {
      "get": {
        "operationId": "tagging_file_v1_tagging_file__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tagging File",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/media": {
      "post": {
        "operationId": "tagging_media_v1_tagging_media_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaggingMediaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAccess"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Media",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/model_input": {
      "post": {
        "operationId": "tagging_model_input_v1_tagging_model_input_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaggingKey"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInputView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Model Input",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/queue": {
      "get": {
        "operationId": "tagging_queue_v1_tagging_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "tier",
            "required": false,
            "schema": {
              "default": "open",
              "maxLength": 32,
              "title": "Tier",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "origin",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 32,
              "title": "Origin",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 2000,
              "maximum": 10000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaggingQueue"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Queue",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/state": {
      "get": {
        "operationId": "tagging_state_v1_tagging_state_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaggingState"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging State",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/suggest": {
      "post": {
        "operationId": "tagging_suggest_v1_tagging_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagSuggestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagSuggestion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Suggest",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/tag": {
      "post": {
        "operationId": "tagging_save_v1_tagging_tag_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagSave"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Save",
        "tags": [
          "tagging"
        ]
      }
    },
    "/v1/tagging/teach": {
      "post": {
        "operationId": "tagging_teach_v1_tagging_teach_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaggingKey"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Tagging Teach",
        "tags": [
          "tagging"
        ]
      }
    }
  }
}